Remove explicit objective constructors
