+++ examples/single_objective.py
@@ -19,9 +19,10 @@ def main():
     main_config = SearchConfiguration(
         generations=5,
         problem_type="min",
+        num_objectives=1,
     )
 
-    tuner = Search.single_objective(
+    tuner = Search(
         objective_function=objective,
         search_space=dna_config,
         search_config=main_config,
